backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
email-validator
psycopg2-binary
pwdlib
pwdlib[argon2]
pwdlib[bcrypt]
PyJWT
python-multipart